chore: move radio, checkbox and option
