+++ packages/checkbox-group/index.js
@@ -1,1 +1,2 @@
 export { LionCheckboxGroup } from './src/LionCheckboxGroup.js';
+export { LionCheckbox } from './src/LionCheckbox.js';
