+++ packages/checkbox-group/lion-checkbox.js
