+++ packages/checkbox-group/src/LionCheckbox.js
