--- packages/checkbox/index.js
@@ -1,1 +1,0 @@
-export { LionCheckbox } from './src/LionCheckbox.js';
