--- packages/checkbox/package.json
@@ -1,43 +1,0 @@
-{
-  "name": "@lion/checkbox",
-  "version": "0.3.16",
-  "description": "A single styleable and accessible checkbox",
-  "author": "ing-bank",
-  "homepage": "https://github.com/ing-bank/lion/",
-  "license": "MIT",
-  "publishConfig": {
-    "access": "public"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/ing-bank/lion.git",
-    "directory": "packages/checkbox"
-  },
-  "scripts": {
-    "prepublishOnly": "../../scripts/npm-prepublish.js"
-  },
-  "keywords": [
-    "lion",
-    "web-components",
-    "checkbox"
-  ],
-  "main": "index.js",
-  "module": "index.js",
-  "files": [
-    "docs",
-    "src",
-    "stories",
-    "test",
-    "translations",
-    "*.js"
-  ],
-  "dependencies": {
-    "@lion/choice-input": "0.7.3",
-    "@lion/core": "0.4.4",
-    "@lion/input": "0.5.16"
-  },
-  "devDependencies": {
-    "@open-wc/demoing-storybook": "^1.10.4",
-    "@open-wc/testing": "^2.5.0"
-  }
-}
