+++ packages/form-system/package.json
@@ -33,7 +33,6 @@
     "*.js"
   ],
   "devDependencies": {
-    "@lion/checkbox": "0.3.16",
     "@lion/checkbox-group": "0.8.1",
     "@lion/core": "0.4.4",
     "@lion/field": "0.11.1",
@@ -46,7 +45,6 @@
     "@lion/input-iban": "0.6.1",
     "@lion/input-range": "0.2.16",
     "@lion/localize": "0.8.8",
-    "@lion/radio": "0.4.6",
     "@lion/radio-group": "0.8.1",
     "@lion/select": "0.5.16",
     "@lion/select-rich": "0.11.3",
