+++ packages/form-system/stories/15-features-overview.stories.mdx
@@ -1,5 +1,6 @@
 import { Story, Meta, html } from '@open-wc/demoing-storybook';
-import '@lion/checkbox/lion-checkbox.js';
+import '@lion/checkbox-group/lion-checkbox-group.js';
+import '@lion/checkbox-group/lion-checkbox.js';
 import '@lion/fieldset/lion-fieldset.js';
 import '@lion/form/lion-form.js';
 import '@lion/input-amount/lion-input-amount.js';
@@ -9,10 +10,10 @@ import '@lion/input-email/lion-input-email.js';
 import '@lion/input-iban/lion-input-iban.js';
 import '@lion/input-range/lion-input-range.js';
 import '@lion/input/lion-input.js';
-import '@lion/option/lion-option.js';
 import '@lion/radio-group/lion-radio-group.js';
-import '@lion/radio/lion-radio.js';
+import '@lion/radio-group/lion-radio.js';
 import '@lion/select/lion-select.js';
+import '@lion/select-rich/lion-option.js';
 import '@lion/select-rich/lion-options.js';
 import '@lion/select-rich/lion-select-rich.js';
 import '@lion/textarea/lion-textarea.js';
@@ -135,7 +136,7 @@ For usage and installation please see the appropriate packages.
 
 ```js
 import '@lion/checkbox-group/lion-checkbox-group.js';
-import '@lion/checkbox/lion-checkbox.js';
+import '@lion/checkbox-group/lion-checkbox.js';
 import '@lion/fieldset/lion-fieldset.js';
 import '@lion/form/lion-form.js';
 import '@lion/input-amount/lion-input-amount.js';
@@ -147,7 +148,7 @@ import '@lion/input-range/lion-input-range.js';
 import '@lion/input/lion-input.js';
 import '@lion/option/lion-option.js';
 import '@lion/radio-group/lion-radio-group.js';
-import '@lion/radio/lion-radio.js';
+import '@lion/radio-group/lion-radio.js';
 import '@lion/select/lion-select.js';
 import '@lion/select-rich/lion-options.js';
 import '@lion/select-rich/lion-select-rich.js';
