--- packages/option/index.js
@@ -1,1 +1,0 @@
-export { LionOption } from './src/LionOption.js';
