--- packages/option/package.json
@@ -1,46 +1,0 @@
-{
-  "name": "@lion/option",
-  "version": "0.5.6",
-  "description": "Allows to provide options for a rich select",
-  "author": "ing-bank",
-  "homepage": "https://github.com/ing-bank/lion/",
-  "license": "MIT",
-  "publishConfig": {
-    "access": "public"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/ing-bank/lion.git",
-    "directory": "packages/option"
-  },
-  "scripts": {
-    "prepublishOnly": "../../scripts/npm-prepublish.js",
-    "test": "cd ../../ && yarn test --grep \"packages/option/test/**/*.test.js\"",
-    "test:watch": "cd ../../ && yarn test:watch --grep \"packages/option/test/**/*.test.js\""
-  },
-  "keywords": [
-    "lion",
-    "web-components",
-    "option"
-  ],
-  "main": "index.js",
-  "module": "index.js",
-  "files": [
-    "docs",
-    "src",
-    "stories",
-    "test",
-    "translations",
-    "*.js"
-  ],
-  "dependencies": {
-    "@lion/choice-input": "0.7.3",
-    "@lion/core": "0.4.4",
-    "@lion/field": "0.11.1"
-  },
-  "devDependencies": {
-    "@open-wc/demoing-storybook": "^1.10.4",
-    "@open-wc/testing": "^2.5.0",
-    "sinon": "^7.2.2"
-  }
-}
