+++ packages/radio-group/index.js
@@ -1,1 +1,2 @@
 export { LionRadioGroup } from './src/LionRadioGroup.js';
+export { LionRadio } from './src/LionRadio.js';
