--- packages/radio/index.js
@@ -1,1 +1,0 @@
-export { LionRadio } from './src/LionRadio.js';
