--- packages/radio/package.json
@@ -1,45 +1,0 @@
-{
-  "name": "@lion/radio",
-  "version": "0.4.6",
-  "description": "Provide a way for users to check a single option amongst a set of choices",
-  "author": "ing-bank",
-  "homepage": "https://github.com/ing-bank/lion/",
-  "license": "MIT",
-  "publishConfig": {
-    "access": "public"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/ing-bank/lion.git",
-    "directory": "packages/radio"
-  },
-  "scripts": {
-    "prepublishOnly": "../../scripts/npm-prepublish.js",
-    "test": "cd ../../ && yarn test --grep \"packages/radio/test/**/*.test.js\"",
-    "test:watch": "cd ../../ && yarn test:watch --grep \"packages/radio/test/**/*.test.js\""
-  },
-  "keywords": [
-    "lion",
-    "web-components",
-    "radio"
-  ],
-  "main": "index.js",
-  "module": "index.js",
-  "files": [
-    "docs",
-    "src",
-    "stories",
-    "test",
-    "translations",
-    "*.js"
-  ],
-  "dependencies": {
-    "@lion/choice-input": "0.7.3",
-    "@lion/core": "0.4.4",
-    "@lion/input": "0.5.16"
-  },
-  "devDependencies": {
-    "@open-wc/demoing-storybook": "^1.10.4",
-    "@open-wc/testing": "^2.5.0"
-  }
-}
