+++ packages/select-rich/index.js
@@ -1,3 +1,4 @@
 export { LionSelectRich } from './src/LionSelectRich.js';
 export { LionSelectInvoker } from './src/LionSelectInvoker.js';
 export { LionOptions } from './src/LionOptions.js';
+export { LionOption } from './src/LionOption.js';
