+++ packages/select-rich/lion-option.js
