+++ packages/select-rich/package.json
@@ -42,13 +42,13 @@
     "@lion/choice-input": "0.7.3",
     "@lion/core": "0.4.4",
     "@lion/field": "0.11.1",
-    "@lion/option": "0.5.6",
     "@lion/overlays": "0.12.3",
     "@lion/validate": "0.7.0"
   },
   "devDependencies": {
     "@lion/form": "0.4.16",
     "@open-wc/demoing-storybook": "^1.10.4",
-    "@open-wc/testing": "^2.5.0"
+    "@open-wc/testing": "^2.5.0",
+    "sinon": "^7.2.2"
   }
 }
