+++ packages/select-rich/src/LionOption.js
