+++ packages/select-rich/stories/index.stories.mdx
@@ -1,6 +1,6 @@
 import { Story, Meta, html, Preview } from '@open-wc/demoing-storybook';
 import { loadDefaultFeedbackMessages, Required } from '@lion/validate';
-import '@lion/option/lion-option.js';
+import '../lion-option.js';
 import '../lion-options.js';
 import '../lion-select-rich.js';
 
@@ -57,7 +57,7 @@ npm i --save @lion/select-rich
 ```js
 import '@lion/select-rich/lion-select-rich.js';
 import '@lion/select-rich/lion-options.js';
-import '@lion/option/lion-option.js';
+import '@lion/select-rich/lion-option.js';
 ```
 
 > No need to npm install `@lion/option` separately, it comes with the rich select as a dependency
